refactor: extract shared personal view permission helpers
