+++ packages/nocodb/src/middlewares/extract-ids/extract-ids.helpers.ts
@@ -1,0 +1,101 @@
+import { ViewLockType } from 'nocodb-sdk';
+
+export const VIEW_KEY = Symbol.for('nc:view');
+
+/**
+ * If the view is a personal view, attach it to the request
+ * so ACL middleware can perform ownership checks.
+ */
+export function markPersonalViewIfNeeded(req: any, view: any) {
+  if (view && view.lock_type === ViewLockType.Personal) {
+    req[VIEW_KEY] = view;
+  }
+}
+
+/**
+ * Check if the request user is the owner of a personal view
+ * attached to the request via VIEW_KEY.
+ */
+export function checkIsPersonalViewOwner(req: any): boolean {
+  return (
+    req[VIEW_KEY]?.lock_type === ViewLockType.Personal &&
+    req[VIEW_KEY].owned_by === req.user?.id
+  );
+}
+
+// Permissions that are view-specific operations checked separately
+// by personalViewOwnerAllowedPermissions. Excluded from the non-owner
+// write restriction check.
+const PERSONAL_VIEW_MANAGEMENT_PERMISSIONS = [
+  'viewColumnUpdate',
+  'viewColumnCreate',
+  'hideAllColumns',
+  'showAllColumns',
+  'gridColumnUpdate',
+  'gridViewUpdate',
+  'galleryViewUpdate',
+  'kanbanViewUpdate',
+  'mapViewUpdate',
+  'calendarViewUpdate',
+  'viewRowColorConditionAdd',
+  'viewRowColorConditionUpdate',
+  'viewRowColorConditionDelete',
+  'viewRowColorSelectAdd',
+  'viewRowColorInfoDelete',
+  'rowColorConditionsFilterCreate',
+] as const;
+
+// Operations excluded from the non-owner personal view write restriction.
+// These are either read operations or view management operations checked
+// separately via personalViewOwnerAllowedPermissions.
+export const viewOperationsExcludedFromPersonalViewCheck = [
+  'viewUpdate',
+  'viewDelete',
+  'dataList',
+  ...PERSONAL_VIEW_MANAGEMENT_PERMISSIONS,
+];
+
+// Permissions that editors can only use on their own personal views.
+// Includes sort/filter CRUD plus view management operations.
+export const editorPersonalViewOnlyPermissions = [
+  'sortCreate',
+  'sortUpdate',
+  'sortDelete',
+  'filterCreate',
+  'filterUpdate',
+  'filterDelete',
+  ...PERSONAL_VIEW_MANAGEMENT_PERMISSIONS,
+];
+
+// All permissions granted to personal view owners regardless of role.
+export const personalViewOwnerAllowedPermissions = [
+  'filterList',
+  'filterGet',
+  'filterChildrenList',
+  'filterCreate',
+  'filterUpdate',
+  'filterDelete',
+  'sortList',
+  'sortGet',
+  'sortCreate',
+  'sortUpdate',
+  'sortDelete',
+  'columnList',
+  'viewUpdate',
+  'viewColumnUpdate',
+  'viewColumnCreate',
+  'hideAllColumns',
+  'showAllColumns',
+  'gridColumnUpdate',
+  'gridViewUpdate',
+  'galleryViewUpdate',
+  'kanbanViewUpdate',
+  'mapViewUpdate',
+  'calendarViewUpdate',
+  'viewRowColorConditionAdd',
+  'viewRowColorConditionUpdate',
+  'viewRowColorConditionDelete',
+  'viewRowColorSelectAdd',
+  'viewRowColorInfoDelete',
+  'rowColorConditionsFilterCreate',
+];
